requirements.txt
Appium-Python-Client==5.2.4
pytest==8.3.4
pytest-html>=4.1.1
python-dotenv>=1.2.1
allure-pytest>=2.15.0
ruff==0.14.6